+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/core/DataConnectAuth.kt
@@ -19,6 +19,7 @@ package com.google.firebase.dataconnect.core
 import com.google.firebase.annotations.DeferredApi
 import com.google.firebase.auth.internal.IdTokenListener
 import com.google.firebase.auth.internal.InternalAuthProvider
+import com.google.firebase.dataconnect.core.DataConnectAuth.GetAuthTokenResult
 import com.google.firebase.dataconnect.core.Globals.toScrubbedAccessToken
 import com.google.firebase.dataconnect.core.LoggerGlobals.debug
 import com.google.firebase.internal.InternalTokenResult
@@ -32,7 +33,7 @@ internal class DataConnectAuth(
   blockingDispatcher: CoroutineDispatcher,
   logger: Logger,
 ) :
-  DataConnectCredentialsTokenManager<InternalAuthProvider>(
+  DataConnectCredentialsTokenManager<InternalAuthProvider, GetAuthTokenResult>(
     deferredProvider = deferredAuthProvider,
     parentCoroutineScope = parentCoroutineScope,
     blockingDispatcher = blockingDispatcher,
@@ -48,7 +49,9 @@ internal class DataConnectAuth(
     provider.removeIdTokenListener(idTokenListener)
 
   override suspend fun getToken(provider: InternalAuthProvider, forceRefresh: Boolean) =
-    provider.getAccessToken(forceRefresh).await().let { GetTokenResult(it.token) }
+    provider.getAccessToken(forceRefresh).await().let { GetAuthTokenResult(it.token) }
+
+  data class GetAuthTokenResult(override val token: String?) : GetTokenResult
 
   private class IdTokenListenerImpl(private val logger: Logger) : IdTokenListener {
     override fun onIdTokenChanged(tokenResult: InternalTokenResult) {