+++ firebase-crashlytics/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 20.0.3
+
 - [fixed] Made creating DataStore files more resilient [#7440]
 
 # 20.0.2